add OnlyCompleteRoots
OnlyCompleteRoots gives a subset of CIDs that can actually be written by a CAR
using only the given blockstore. We're running into an issue feeding offline-only
blockstores to WriteCar that have one or more incomplete / corrupted DAGs, causing
the CAR to fail too early.

Because car files write root info to the backing writer immideately, a check like this
has to happen before acrhive writing begins. Most users won't want to do this because
it adds an extra pass of reads to the entire store, so this API presents
OnlyCompleteRoots as an optional function.